chore(Test): Added sinon for events testing.
