+++ package.json
@@ -57,6 +57,7 @@
     "prettier": "^2.2.1",
     "rollup-plugin-postcss": "^4.0.0",
     "sass": "^1.35.1",
+    "sinon": "^11.1.1",
     "tslib": "^2.2.0",
     "typescript": "^4.3.2"
   },
